Rename and remove some unused code
