+++ maisim/maisim.Game/Graphics/UserInterface/MaisimButton.cs
@@ -1,10 +1,8 @@
 using maisim.Game.Graphics.Sprites;
 using osu.Framework.Allocation;
-using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
 using osuTK;
 using osuTK.Graphics;
@@ -13,9 +11,6 @@ namespace maisim.Game.Graphics.UserInterface
 {
     public class MaisimButton : Button
     {
-        private readonly Color4 buttonColor;
-        private readonly Color4 buttonOutlineColor;
-
         [BackgroundDependencyLoader]
         private void load()
         {
@@ -24,9 +19,6 @@ namespace maisim.Game.Graphics.UserInterface
 
         public MaisimButton(string text, Color4 buttonColor, Color4 buttonOutlineColor)
         {
-            this.buttonColor = buttonColor;
-            this.buttonOutlineColor = buttonOutlineColor;
-
             InternalChild = new Container
             {
                 Anchor = Anchor.Centre,
@@ -34,6 +26,8 @@ namespace maisim.Game.Graphics.UserInterface
                 Size = new Vector2(350, 50),
                 Masking = true,
                 CornerRadius = 30,
+                BorderColour = buttonOutlineColor,
+                BorderThickness = 5,
                 Children = new Drawable[]
                 {
                     new Box
